mk/subst.mk: document the unspecified application order
To avoid bmake warnings because of duplicate class names, the :O:u
modifier had been added in r1.66 on 2020-03-21.  This had the side effect
that the subst classes are now applied in alphabetical order instead of
declaration order.

For this to actually matter, there must be a file that is affected by two
different subst classes and in which the substitutions depend on each
other or prevent each other.  Chances for that are pretty low.

The order is intentionally documented as being unspecified, to allow for
future modifications, just in case that a bmake variable modifier is
invented that filters for duplicates without requiring the duplicates to
be adjacent to each other.  In that situation, it would be nicer to
switch back to declaration order instead of alphabetical.