www/ruby-faraday_middleware: update to 1.0.0
Update ruby-faraday_middleware to 1.0.0.

pkgsrc change: add "USE_LANGUAGES=	# none".


1.0.0 (2020-03-29)
This is FaradayMiddleware 1.0 official release!

This release officially supports Faraday 1.0.
It contains all changes included in the 1.0.0.rc1 release, plus:

* Increase minimum rake version to 12.3.3 for security reasons

Special thanks to contributors who helped us push this over the line:
@BobbyMcWho, @onk and @d-m-u


1.0.0.rc1 (2020-02-28)

This is a prerelease. See #200 for details on the shipping plans for 1.0.0.

Changes

* Support Ruby 2.3+.
* Drop support for Faraday v0.7

Bugs fixed

* #198 Autoload an exception class correctly

Chores and misc

* Lint the codebase with RuboCop
* Closed the Wiki and turned it into Markdown files in docs/


0.14.0 (2020-01-10)

NOTE: This is the last minor release in the v0.x series, next release will
be 1.0 to match Faraday v1.0 release and from then on only fixes will be
applied to v0.14.x!

* Fix support for empty bodies having Content-Encoding Gzip (#192)
* RSpec: disable_monkey_patching (#191)
* Add ResponseMiddleware.define_parser unit test (#190)
* Use full URL (url.host + url.request_uri) as cache key (#195)


0.13.1 (2019-02-04)

* Clear Authorization header when redirecting cross-site (#183)
* Brotli is an optional dependency for Gzip (#186)


0.13.0 (2019-01-22)

* Adds support for Brotli decompression of responses (#134)
* Adds support for JSON 2 (#169)
* Hash request URIs before using them as a cache key (#181)
* Remove Addressable::URI query hack (#184)