py-hypothesis: updated to 4.57.1
4.57.1:
This patch improves the type hints and documentation for the django extra. There is no runtime change.

4.57.0:
This release improves support for the SupportsOp protocols from the typing module when using on from_type() as outlined in issue 2292. The following types now generate much more varied strategies when called with from_type():

typing.SupportsAbs
typing.SupportsBytes
typing.SupportsComplex
typing.SupportsInt
typing.SupportsFloat
typing.SupportsRound
Note that using from_type() with one of the above strategies will not ensure that the the specified function will execute successfully (ie : the strategy returned for from_type(typing.SupportsAbs) may include NaNs or things which cause the abs() function to error. )

4.56.3:
This release fixes a small internal bug in shrinking which could have caused it to perform slightly more tests than were necessary. Fixing this shouldn’t have much effect but it will make shrinking slightly faster.

4.56.2:
This release removes an internal heuristic that was no longer providing much benefit. It is unlikely that there will be any user visible effect.

4.56.1:
This release further improves the optimisation algorithm for targeted property-based testing.

4.56.0:
This release enables deprecation warnings even when the verbosity setting is quiet, in preparation for Hypothesis 5.0.

Warnings can still be filtered by the standard mechanisms provided in the standard-library warnings module.

4.55.4:
This release improves Hypothesis’s management of the set of test cases it tracks between runs. It will only do anything if you have Phase.target enabled and an example database set. In those circumstances it should result in a more thorough and faster set of examples that are tried on each run.

4.55.3:
This release makes Hypothesis better at generating test cases where generated values are duplicated in different parts of the test case. This will be especially noticeable with reasonably complex values, as it was already able to do this for simpler ones such as integers or floats.

4.55.2:
This release expands the set of test cases that Hypothesis saves in its database for future runs to include a representative set of “structurally different” test cases - e.g. it might try to save test cases where a given list is empty or not.

Currently this is unlikely to have much user visible impact except to produce slightly more consistent behaviour between consecutive runs of a test suite. It is mostly groundwork for future improvements which will exploit this functionality more effectively.

4.55.1:
This patch fixes issue 2257, where from_type() could incorrectly generate bytestrings when passed a generic typing.Sequence such as Sequence[set].

4.55.0:
This release adds database support for targeted property-based testing, so the best examples based on the targeting will be saved and reused between runs. This is mostly laying groundwork for future features in this area, but will also make targeted property-based tests more useful during development, where the same tests tend to get run over and over again.

If max_examples is large, this may increase memory usage significantly under some circumstances, but these should be relatively rare.

This release also adds a dependency on the sortedcontainers package.

4.54.2:
This release improves the optimisation algorithm for targeted property-based testing, so that it will find higher quality results more reliably. Specifically, in cases where it would previously have got near a local optimum, it will now tend to achieve the locally optimal value.

4.54.1:
This release is mostly internal changes in support of better testing of the core engine. You are unlikely to see much effect, although some internal heuristics have changed slightly.

4.54.0:
This release adds a dedicated phase for targeted property-based testing, and (somewhat) improves the targeting algorithm so that it will find higher quality results more reliably. This comes at a cost of making it more likely to get stuck in a local optimum.